targetable: Namespace Core API kind wrapper
